Add CSV export
This using manual writes, falling back on JSON rendering for some complicated
fields. It was tempting to try this by contorting Zeek's ASCII logs but the
manual route provides more flexibility, such as optionally writing a custom
header.

Escaping may need iteration and is pretty basic right now, just looking for '\n'
in the JSON-escaped strings and replacing with '""'.