Expand configurability of Zeek package determination for log fields
There's now a vector `Log::Schema::package_prefixes` that configures the paths
the package considers as lead-up to a package directory. The default of
considering the "site/packages" directory as such a location remains unchanged.

The determination of these paths now happens after script path normalization.

The documentation now explains how to customize this, all the way to fully
overriding package names in an adapt hook handler.

Includes test cases.